pyproject.toml
[project]
name = "ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.123.9",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.235",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.0.0",
    "opencv-python>=4.8.0",
    "timm>=0.9.0",
    "numpy<2",
    "pillow-heif>=0.13.0",
    "matplotlib>=3.8.0",
    "pandas>=2.3.3",
]

[tool.uv]
required-environments = ["sys_platform == 'linux' and platform_machine == 'x86_64'", "sys_platform == 'darwin'"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
